rag_elastic_local/requirements.txt
elasticsearch==9.0.0
fastapi==0.115.12
pandas==2.2.3
pdfplumber==0.11.6
python-dotenv==1.1.0
python_pptx==1.0.2
tqdm==4.67.1
uvicorn==0.34.1
google-generativeai==0.8.4
ollama==0.4.7